finish restructure
