+++ src/Catalyst.Cli.IntegrationTests/Commands/AddFileCommandTest.cs
@@ -34,7 +34,7 @@ using Xunit.Abstractions;
 
 namespace Catalyst.Cli.IntegrationTests.Commands
 {
-    public sealed class AddFileCommandTest : CliCommandsTests
+    public sealed class AddFileCommandTest : CliCommandTestBase
     {
         //This test is the base to all other tests.  If the Cli cannot connect to a node than all other commands
         //will fail
