+++ src/Catalyst.Cli.IntegrationTests/Commands/CliCommandTestBase.cs
@@ -27,23 +27,17 @@ using System.IO;
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using Autofac;
-using Catalyst.Common.Extensions;
-using Catalyst.Common.Interfaces.Cli;
 using Catalyst.Common.Interfaces.Rpc;
 using Catalyst.Common.UnitTests.TestUtils;
-using Catalyst.Protocol.Common;
-using Catalyst.Protocol.Rpc.Node;
 using DotNetty.Transport.Channels;
-using FluentAssertions;
 using Microsoft.Extensions.Configuration;
 using NSubstitute;
-using Xunit;
 using Xunit.Abstractions;
 using Constants = Catalyst.Common.Config.Constants;
 
 namespace Catalyst.Cli.IntegrationTests.Commands
 {
-    public class CliCommandsTests : ConfigFileBasedTest
+    public abstract class CliCommandTestBase : ConfigFileBasedTest
     {
         protected readonly INodeRpcClient NodeRpcClient;
 
@@ -60,7 +54,7 @@ namespace Catalyst.Cli.IntegrationTests.Commands
                 new object[] {"/fake_file_hash", AppDomain.CurrentDomain.BaseDirectory + "/Config/addfile_test.json", true}
             };
 
-        public CliCommandsTests(ITestOutputHelper output) : base(output)
+        public CliCommandTestBase(ITestOutputHelper output) : base(output)
         {
             var config = new ConfigurationBuilder()
                .AddJsonFile(Path.Combine(Constants.ConfigSubFolder, Constants.ShellComponentsJsonConfigFile))
@@ -84,37 +78,6 @@ namespace Catalyst.Cli.IntegrationTests.Commands
             ConfigureContainerBuilder(config);
 
             ContainerBuilder.RegisterInstance(nodeRpcClientFactory).As<INodeRpcClientFactory>();
-        }
-
-        [Fact]
-        public void Cli_Can_Connect_To_Node()
-        {
-            using (var container = ContainerBuilder.Build())
-            {
-                using (container.BeginLifetimeScope(CurrentTestName))
-                {
-                    var shell = container.Resolve<ICatalystCli>();
-                    var hasConnected = shell.AdvancedShell.ParseCommand("connect", "-n", "node1");
-                    hasConnected.Should().BeTrue();
-                }   
-            }
-        }
-        
-        [Fact]
-        public void Cli_Can_Handle_Multiple_Connection_Attempts()
-        {
-            using (var container = ContainerBuilder.Build())
-            {
-                using (container.BeginLifetimeScope(CurrentTestName))
-                {
-                    var shell = container.Resolve<ICatalystCli>();
-                    for (var i = 0; i < 10; i++)
-                    {
-                        var canConnect = shell.AdvancedShell.ParseCommand("connect", "-n", "node1");
-                        canConnect.Should().BeTrue();
-                    }
-                }   
-            }
         }
     }
 }
