+++ src/Catalyst.Cli.IntegrationTests/Commands/GetFileCommandTest.cs
@@ -34,7 +34,7 @@ using Xunit.Abstractions;
 
 namespace Catalyst.Cli.IntegrationTests.Commands
 {
-    public sealed class GetFileCommandTest : CliCommandsTests
+    public sealed class GetFileCommandTest : CliCommandTestBase
     {
         //This test is the base to all other tests.  If the Cli cannot connect to a node than all other commands
         //will fail
