+++ src/Catalyst.Cli.IntegrationTests/Commands/GetMempoolTest.cs
@@ -30,7 +30,7 @@ using Xunit.Abstractions;
 
 namespace Catalyst.Cli.IntegrationTests.Commands
 {
-    public sealed class GetMempoolCommandTest : CliCommandsTests
+    public sealed class GetMempoolCommandTest : CliCommandTestBase
     {
         //This test is the base to all other tests.  If the Cli cannot connect to a node than all other commands
         //will fail
