+++ src/Catalyst.Cli.IntegrationTests/Commands/PeerRemoveCommandTest.cs
@@ -34,7 +34,7 @@ using Xunit.Abstractions;
 
 namespace Catalyst.Cli.IntegrationTests.Commands
 {
-    public sealed class PeerRemoveCommandTest : CliCommandsTests
+    public sealed class PeerRemoveCommandTest : CliCommandTestBase
     {
         //This test is the base to all other tests.  If the Cli cannot connect to a node than all other commands
         //will fail
